requirements.txt
Flask
numpy
Pillow
torch
torchvision
